requirements.txt
flask==3.0.0
flask-sqlalchemy==3.1.1
requests==2.31.0
python-dotenv==1.0.0
pytest==7.4.3
gunicorn==21.2.0
prometheus_flask_exporter==0.22.4
python-json-logger==2.0.7
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-instrumentation-flask==0.43b0
opentelemetry-exporter-otlp==1.22.0